Rename ThreadNetworkDirectory files for SCI conversion
